UHF-9712: Implement OGImageBuilderInterface
